+++ beemgraphenebase/__init__.py
@@ -9,6 +9,7 @@ from .version import version as __version__
 
 __all__ = ['account',
            'base58',
+           'bip32'
            'bip38',
            'transactions',
            'types',